[refactor]チャートコンポーネントのロジックをcomposablesに切り出す
