+++ src/components/ChartSection.vue
@@ -2,7 +2,7 @@
   <section class="chart-card">
     <div class="type">
       <label
-        v-for="(label, index) in labelList"
+        v-for="(label, index) in chartUtils.labelList"
         :key="index"
         class="chart-radio"
         :for="label"
@@ -10,7 +10,7 @@
         {{ label }}
         <input
           :id="label"
-          v-model="selectedLabel"
+          v-model="chartUtils.selectedLabel.value"
           type="radio"
           :value="label"
         />
@@ -20,103 +20,25 @@
       width="100%"
       type="line"
       height="400"
-      :options="chartOptions"
-      :series="series"
+      :options="chartUtils.chartOptions.value"
+      :series="chartUtils.series.value"
     ></apexchart>
   </section>
 </template>
 <script lang="ts" setup>
-import {
-  PopulationForChart,
-  Population,
-  PopulationYearAndValue,
-  Prefecture,
-} from '~/composables/state';
-const prefectureState = usePrefecture();
-const populationState = useChart();
-const api = useResasApi();
 const chartUtils = useChartUtils();
-const labelList = chartUtils.labelList;
-const selectedLabel = ref(labelList[0]); // 総人口
-const findPopulation = (populationForChart: PopulationForChart) => {
-  if (
-    !populationForChart ||
-    !populationForChart.data ||
-    populationForChart.data.length === 0
-  )
-    return;
-  const population = populationForChart.data.find(
-    (population: Population) => population.label === selectedLabel.value
-  );
-  return population;
-};
-const series = ref(chartUtils.initializeSeries());
-const chartOptions = ref(chartUtils.initializeOptions());
-/**
- * チェックした都道府県の人口構成を1データずつ整形し、チャートにセットする
- */
-const setPopulationToSeries = () => {
-  series.value = populationState.list.value.map(
-    (population: PopulationForChart) => ({
-      name: prefectureState.list.value.find(
-        (prefecture: Prefecture) => prefecture.prefCode === population.id
-      ).prefName,
-      id: population.id,
-      data: findPopulation(population)
-        ? findPopulation(population).data.map(
-            (populationYearAndValue: PopulationYearAndValue) =>
-              populationYearAndValue.value
-          )
-        : [],
-    })
-  );
-};
-const updatePopulation = async (
-  newCheck: Array<number>,
-  oldCheck: Array<number>
-) => {
-  if (oldCheck && newCheck.length < oldCheck.length) {
-    populationState.list.value.forEach(
-      (populationForChart: PopulationForChart, index: number) => {
-        if (!newCheck.includes(populationForChart.id)) {
-          populationState.splicePopulation(index);
-        }
-      }
-    );
-  } else {
-    populationState.pushPopulation({
-      id: newCheck[newCheck.length - 1],
-      ...(await api.getPopulationList({
-        prefCode: newCheck[newCheck.length - 1],
-      })),
-    });
-  }
-  setPopulationToSeries();
-  chartOptions.value = {
-    ...chartOptions.value,
-    xaxis: {
-      categories: findPopulation(populationState.list.value[0])
-        ? findPopulation(populationState.list.value[0]).data.map(
-            (populationYearAndValue: PopulationYearAndValue) =>
-              populationYearAndValue.year
-          )
-        : [],
-      title: {
-        text: '年',
-      },
-    },
-  };
-};
+const prefectureState = usePrefecture();
+
 watch(
   prefectureState.checkedPrefecture,
   // 秒間リクエストに制限があるため、debounceをかける
   useDebounce(async (newCheck: Array<number>, oldCheck: Array<number>) => {
-    await updatePopulation(newCheck, oldCheck);
+    await chartUtils.updatePopulation(newCheck, oldCheck);
   }, 200)
 );
 
-watch(selectedLabel, () => {
-  setPopulationToSeries();
+watch(chartUtils.selectedLabel, () => {
+  chartUtils.setPopulationToSeries();
 });
 </script>
